+++ ReelChoice/settings.py
@@ -122,3 +122,6 @@ STATIC_URL = 'static/'
 # https://docs.djangoproject.com/en/5.2/ref/settings/#default-auto-field
 
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
+
+LOGIN_REDIRECT_URL = "reelchoice_app:home"
+LOGOUT_REDIRECT_URL = "reelchoice_app:login"
