+++ ReelChoice/urls.py
@@ -15,8 +15,12 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path
+from django.urls import path, include
+from django.conf import settings
+from django.conf.urls.static import static
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-]
+    path("", include(("reelchoice_app.urls", "reelchoice_app"), "reelchoice_app"))
+] + static(settings.STATIC_URL)
+
