+++ reelchoice_app/static/templates/base.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+ <meta charset="UTF-8">
+ <meta name="viewport" content="width=device-width, initial-scale=1.0">
+ <title>ReelChoice</title>
+ <link rel="preconnect" href="https://fonts.googleapis.com">
+<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+<link href="https://fonts.googleapis.com/css2?family=Mulish:wght@400;700&display=swap" rel="stylesheet">
+ {% load static %}
+ <link rel="stylesheet" href="{% static "css/style.css" %}">
+</head>
+<body>
+  {% block content %}{% endblock content %}
+</body>
+</html>
