+++ reelchoice_app/static/templates/home.html
