+++ reelchoice_app/templates/registration/login.html
@@ -1,0 +1,14 @@
+{% extends "base.html" %}
+{% block content %}
+<div class="shadow-wrap">
+ <div class="form-wrap">
+  <h3>Login </h3>
+<form method="POST">
+ {% csrf_token %}
+ {{form}}
+ <button type="submit">Submit</button>
+</form>
+<a class="signup-link" href="{% url "reelchoice_app:authView" %}">Signup</a>
+</div>
+</div>
+{% endblock content %}a
