+++ reelchoice_app/templates/registration/signup.html
@@ -1,0 +1,14 @@
+{% extends "base.html" %}
+{% block content %}
+<div class="shadow-wrap">
+ <div class="form-wrap">
+  <h3>Signup </h3>
+<form method="POST">
+ {% csrf_token %}
+ {{form}}
+ <button type="submit">Submit</button>
+</form>
+<a class="signup-link" href="{% url "reelchoice_app:login" %}">Login</a>
+</div>
+</div>
+{% endblock content %}a
