+++ reelchoice_app/views.py
@@ -1,3 +1,19 @@
-from django.shortcuts import render
+from django.shortcuts import render, redirect
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.decorators import login_required
 
-# Create your views here.
+
+@login_required
+def home(request):
+ return render(request, "home.html", {})
+
+
+def authView(request):
+ if request.method == "POST":
+  form = UserCreationForm(request.POST or None)
+  if form.is_valid():
+   form.save()
+   return redirect("reelchoice_app:login")
+ else:
+  form = UserCreationForm()
+ return render(request, "registration/signup.html", {"form": form})
